Enable installation with configure defaults
The configure default for the man page is to not generate it. However,
this default results in an error during installation.

Add a pre-generated man page.

Check the prerequisites to regenerate the man page during configuration
instead of reporting an error later during the build process, and remove
the pre-generated man page after checking, so that --enable-man will
regenerate the man page no matter what.